Split merged_prs and release stats report in two separate commands
Make them print to stdout instead of writing to a file
