+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/stats/__init__.py
@@ -4,9 +4,9 @@
 import click
 
 from ...console import CONTEXT_SETTINGS
-from .csv_report import csv_report
+from .stats import report,merged_prs
 
-ALL_COMMANDS = (csv_report,)
+ALL_COMMANDS = (report,merged_prs)
 
 
 @click.group(context_settings=CONTEXT_SETTINGS, short_help='A collection of tasks to generate reports about releases')
